Simplify card selection and prepare to fix safari cursor issue.
